+++ lib/src/screens/image/widgets/interactive_viewer_with_scale.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:flutter/widgets.dart';
-import 'package:privacyblur/src/screens/image/helpers/filter_position.dart';
+import 'package:privacyblur/src/screens/image/utils/filter_position.dart';
 import 'package:privacyblur/src/screens/image/helpers/image_states.dart';
 import 'package:privacyblur/src/utils/image_filter/helpers/filter_result.dart';
 
@@ -147,7 +147,7 @@ class _InteractiveViewerWithScaleState
     widget.extState.positions.asMap().forEach((key, value) {
       var tmp =
           sqrt(pow(value.posX - offset.dx, 2) + pow(value.posY - offset.dy, 2));
-      if (value.isInnerPoint(offset.dx.toInt(), offset.dy.toInt())) {
+      if (value.isInnerPoint(offset.dx, offset.dy)) {
         if (tmp < dist) {
           index = key;
           dist = tmp;
